requirements.txt
jwt
dash
pandas
Flask
Flask-Assets
Flask-RESTful
Flask-Migrate
Flask-SQLAlchemy
flasgger
botocore
psycopg2-binary
botocore
sqlalchemy-cockroachdb
opentelemetry-sdk-extension-aws
opentelemetry-propagator-aws-xray
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp
opentelemetry-instrumentation-botocore
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-dbapi
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
pytest
gunicorn